FLOAT

+++ backend/db/models/event.js
@@ -52,7 +52,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     price: {
-      type: DataTypes.DECIMAL(10,2),
+      type: DataTypes.FLOAT(10,2),
       allowNull: false,
     },
     startDate: {
